further int tweaks for py3.10+ compat

+++ PYME/Acquire/ui/intsliders.py
@@ -53,7 +53,7 @@ class IntegrationSliders_(wx.Panel):
         
         for c in range(nsliders):
             #if not sys.platform == 'darwin':
-            sl = wx.Slider(self, -1, self.chaninfo.itimes[c], 1, min(5*self.chaninfo.itimes[c], 10000), size=wx.Size(100,-1),style=wx.SL_HORIZONTAL|wx.SL_AUTOTICKS)#|wx.SL_LABELS)
+            sl = wx.Slider(self, -1, int(self.chaninfo.itimes[c]), 1, int(min(5*self.chaninfo.itimes[c], 10000)), size=wx.Size(100,-1),style=wx.SL_HORIZONTAL|wx.SL_AUTOTICKS)#|wx.SL_LABELS)
             #else:#workaround for broken mouse event handling (and hence sliders) on MacOS
             #    sl = wx.Slider(self, -1, self.chaninfo.itimes[c], 1, min(5*self.chaninfo.itimes[c], 10000), size=wx.Size(300,-1),style=wx.SL_HORIZONTAL|wx.SL_AUTOTICKS|wx.SL_LABELS)
 
@@ -142,7 +142,7 @@ class IntegrationSliders(wx.Panel):
         itime = 1e3*self.scope.state['Camera.IntegrationTime']
 
         
-        sl = wx.Slider(self, -1, itime, 1, min(5*itime, 10000), size=wx.Size(100,-1),style=wx.SL_HORIZONTAL)#|wx.SL_AUTOTICKS)#|wx.SL_LABELS)
+        sl = wx.Slider(self, -1, int(itime), 1, int(min(5*itime, 10000)), size=wx.Size(100,-1),style=wx.SL_HORIZONTAL)#|wx.SL_AUTOTICKS)#|wx.SL_LABELS)
         
         sl_val = wx.ComboBox(self, -1, choices = timeChoices, value = '%d' % itime, size=(65, -1), style=wx.CB_DROPDOWN|wx.TE_PROCESS_ENTER)
